pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "outboundai"
version = "2.0.0"
description = "HR Tech Lead Generation System"
authors = [{name = "Ariel", email = "ariel@cliocircle.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.5.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "python-dotenv>=0.19.0",
    "langchain-nvidia-ai-endpoints>=0.1.0",
    "langchain-core>=0.1.0",
    "serpapi>=0.1.0",
    "pyhunter>=0.1.0",
    "google-auth-oauthlib>=0.5.0",
    "google-api-python-client>=2.0.0",
    "ratelimit>=2.2.0",
    "pyyaml>=6.0",
    "apify-client>=1.0.0",
    "pdfplumber>=0.7.0",
    "dateutil>=2.8.0",
    "certifi>=2022.0.0",
    "urllib3>=1.26.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "isort>=5.10.0",
    "mypy>=0.991",
    "bandit>=1.7.0",
    "safety>=2.0.0"
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503", "E501"]
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "build",
    "dist",
    "*.egg-info"
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml"
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B601"]